requirements.txt
tensorflow-gpu
tensorflow-probability
numpy
matplotlib
tqdm
scikit-learn
clingo
tqdm